Kmbox Config stuff
This version changes the LCD Screen so that everytime you run the program it shows that it is active. It also reboots the Kmbox upon close to ensure safety of the box at exit. If the LCD screen does not change, simply umplug and plug the box back in. It should change however if you exit the program safely with Alt + F4, Exit App, or Home+End hotkey.