Kmbox security Update
This update pushes out the Kmbox being connected or not through a single ping of data on launch and when pressing connect. Also displays clear text on the menu next to the Kmbox details if the connection is successful or not successful. 

Any concerns or issues, please open an Issue tab, or contact me directly on discord at Warlockmnk.

Enjoy!